requirements.txt
streamlit
langchain-groq
Pillow
requests